Add files via upload
This HTML page serves as an invitation to a birthday celebration, providing essential details such as the date (24th of August), accompanied by a vibrant image of a birthday cake. It includes a list of recommended items for guests to bring, emphasizing the host's love for balloons and their appetite for cake. Additionally, there's a link provided for attendees to easily access the location on Google Maps, ensuring they can find their way to the event effortlessly.